improved packing heuristic
At the end of each execution phase, there will be a brief period when there are not enough jobs to fill all available threads.  The length of that period is determined by the length of the jobs that are last in the execution order; by saving the shortest jobs for last, we significantly improve occupancy.