grammar refactor: use one map
I'm not sure whether this is better or worse than the current approach.
During refactoring I found a lot more hashmap iterations than I
expected, and iteration over just the constants or just the variables is
worse in this implementation than with two maps.
On the other hand, it decreases the memory footprint of each node
in half, and there is less internal fragmentation in the maps.
@tirix, could you test this to make sure it still works and performs
alright?