Begin adding the ability to verify definitions
This adds a --verify-definitions 'globlist' option
to the set of options and starting code to implement it.

This version doesn't actually perform the check, it instead
simply prints as the names of the statements to check.
More importantly, it begins setting up the code infrastructure
so we can implement the definitions check "for real".

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>