Fix set.mm check
First of all, use the develop branch rather than the master branch
because the latter is rather old and we haven't been using it.

Second of all, add --verify-markup and --parse-typesetting so that we
run more of the checks (including most notably the ones which the set.mm
automated checks enable).